project-1/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
passlib[bcrypt]
python-jose
python-multipart
jinja2
